print.multcompLetters {multcompView}	R Documentation

print a multcompLetters object

Description

print method for an object of class 'multcompLetters'.

Usage


## S3 method for class 'multcompLetters'
print(x, all = FALSE, ...)

Arguments

x
an object of class 'multcompLetters'

all
FALSE to print only the character vector representations of the 'multcompLetters' comparison summary; TRUE to print also the matrix representation.

...
Other optional print parameters as described on the print help page.

Details

Prints only the Letters component of the 'multcompLetters' list unless all=TRUE.

Value

the named, character vector representation of the 'multcompLetters' evaluation of the distance relationships

Author(s)

Spencer Graves

See Also

multcompLetters

Examples



dif3 <- c(FALSE, FALSE, TRUE)
names(dif3) <- c("A-B", "A-C", "B-C")
dif3L <- multcompLetters(dif3)
dif3L
print(dif3L)
print(dif3L, TRUE)


[Package multcompView version 0.1-10 Index]