vec2mat2 {multcompView}	R Documentation

Convert a vector of hyphenated names into a character matrix.

Description

Convert a vector of hyphenated names into a character matrix with 2 columns containing the names split in each row.

Usage


vec2mat2(x, sep = "-")

Arguments

x
Vector of hyphenated names

sep
"strsplit" character to apply to names(x).

Details

If each element of x does not contain exactly 1 "sep" character, an error is issued.

Value

A character matrix with rownames = x and with the character string preceding the "sep" character in the first column and the character string following the "sep" character in the second column.

Author(s)

Spencer Graves

See Also

vec2mat multcompLetters

Examples



vec2mat2(c("a-b", "a-c", "b-c"))

vec2mat2(c("a-b", "b-a"))

 


[Package multcompView version 0.1-10 Index]